auto_creative_engine/requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
pillow>=10.0.0

# Google Gemini API
google-generativeai>=0.3.0

# Image processing
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
